requirements.txt
typer>=0.9.0
aiohttp>=3.9.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aioresponses>=0.7.0
jinja2>=3.1.2
pyyaml>=6.0
# Chart.js is embedded in the HTML report via CDN — no pip package needed